Make commonconfig.DeploymentConfig private configuration in the observation service
