+++ observation-service/config/config_test.go
@@ -17,7 +17,7 @@ func TestDefaultConfigs(t *testing.T) {
 	emptyStringMap := make(map[string]string)
 	defaultCfg := Config{
 		Port: 9001,
-		DeploymentConfig: commonconfig.DeploymentConfig{
+		DeploymentConfig: DeploymentConfig{
 			EnvironmentType: "local",
 			LogLevel:        commonconfig.InfoLevel,
 			MaxGoRoutines:   1000,
@@ -91,7 +91,7 @@ func TestLoadConfigFiles(t *testing.T) {
 			configFiles: []string{"../testdata/config1.yaml", "../testdata/config2.yaml"},
 			expected: Config{
 				Port: 9002,
-				DeploymentConfig: commonconfig.DeploymentConfig{
+				DeploymentConfig: DeploymentConfig{
 					EnvironmentType: "dev",
 					LogLevel:        commonconfig.InfoLevel,
 					MaxGoRoutines:   1000,
