+++ observation-service/logger/observation_logger_test.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	commonconfig "github.com/caraml-dev/timber/common/config"
 	"github.com/caraml-dev/timber/observation-service/config"
 	"github.com/caraml-dev/timber/observation-service/services"
 )
@@ -15,7 +14,7 @@ func TestObservationLogger(t *testing.T) {
 	// Configs
 	consumerConfig := config.LogConsumerConfig{}
 	producerConfig := config.LogProducerConfig{}
-	deploymentConfig := commonconfig.DeploymentConfig{}
+	deploymentConfig := config.DeploymentConfig{}
 	metricConfig := config.MonitoringConfig{}
 	metricService, err := services.NewMetricService(deploymentConfig, metricConfig)
 	assert.NoError(t, nil, err)
