+++ observation-service/services/metric_service_test.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/stretchr/testify/suite"
 
-	commonconfig "github.com/caraml-dev/timber/common/config"
 	"github.com/caraml-dev/timber/observation-service/config"
 	"github.com/caraml-dev/timber/observation-service/internal/testutils"
 	"github.com/caraml-dev/timber/observation-service/monitoring"
@@ -24,7 +23,7 @@ func (s *MetricServiceTestSuite) SetupSuite() {
 
 	var err error
 	s.cfg = config.Config{
-		DeploymentConfig: commonconfig.DeploymentConfig{
+		DeploymentConfig: config.DeploymentConfig{
 			ProjectName: "default",
 			ServiceName: "observation-service",
 		},
